+++ examples/express/webpack.config.js
@@ -1,17 +1,27 @@
 import { resolve, basename, dirname, posix } from "path";
-import externals from "webpack-node-externals";
-import { TogglePointInjection } from "@asos/web-toggle-point-webpack/plugins";
-import MiniCssExtractPlugin from "mini-css-extract-plugin";
 import { fileURLToPath } from "url";
+import externals from "webpack-node-externals";
+import { TogglePointInjectionPlugin } from "@asos/web-toggle-point-webpack/plugins";
+import MiniCssExtractPlugin from "mini-css-extract-plugin";
+import staticLoadStrategyFactory from "@asos/web-toggle-point-webpack/moduleLoadStrategyFactories/staticLoadStrategyFactory";
+import lazyComponentLoadStrategyFactory from "@asos/web-toggle-point-react-pointcuts/lazyComponentLoadStrategyFactory";
 
 const configPointCutConfig = {
   name: "configuration variants",
   variantGlob: "./src/routes/config/__variants__/*/*/*.jsx",
-  togglePointModule: "/src/routes/config/togglePoint.js"
+  togglePointModuleSpecifier: "/src/routes/config/togglePoint.js",
+  loadStrategy: lazyComponentLoadStrategyFactory()
 };
 
 const common = {
   mode: "production",
+  devtool: "source-map",
+  experiments: {
+    outputModule: true
+  },
+  output: {
+    module: true
+  },
   module: {
     rules: [
       {
@@ -26,24 +36,25 @@ const common = {
         use: [MiniCssExtractPlugin.loader, "css-loader"]
       }
     ]
-  }
+  },
+  plugins: [new MiniCssExtractPlugin()]
 };
 
 const config = [
   {
     entry: "./src/index.js",
     target: "node",
+    ...common,
     output: {
+      ...common.output,
       path: resolve(dirname(fileURLToPath(import.meta.url)), "bin"),
-      filename: "server.cjs",
-      clean: true,
-      chunkFormat: "module"
+      filename: "server.mjs",
+      clean: true
     },
     externals: [externals()],
-    ...common,
     plugins: [
-      new MiniCssExtractPlugin(),
-      new TogglePointInjection({
+      ...common.plugins,
+      new TogglePointInjectionPlugin({
         pointCuts: [
           configPointCutConfig,
           {
@@ -55,7 +66,8 @@ const config = [
                 ...Array(3).fill(".."),
                 basename(variantPath)
               ),
-            togglePointModule: "/src/routes/animals/togglePoint.js"
+            togglePointModuleSpecifier: "/src/routes/animals/togglePoint.js",
+            loadStrategy: staticLoadStrategyFactory()
           }
         ]
       })
@@ -64,15 +76,27 @@ const config = [
   {
     entry: "./src/routes/config/client.js",
     target: "web",
+    ...common,
     output: {
+      ...common.output,
       path: resolve(dirname(fileURLToPath(import.meta.url)), "public"),
-      filename: "main.js"
+      filename: "main.mjs"
     },
     plugins: [
-      new MiniCssExtractPlugin(),
-      new TogglePointInjection({ pointCuts: [configPointCutConfig] })
+      ...common.plugins,
+      new TogglePointInjectionPlugin({ pointCuts: [configPointCutConfig] })
     ],
-    ...common
+    module: {
+      ...common.module,
+      rules: [
+        {
+          test: /\.js$/,
+          enforce: "pre",
+          use: ["source-map-loader"]
+        },
+        ...common.module.rules
+      ]
+    }
   }
 ];